// LDC.Atlas.WebApp/Client/app/shared/entities/default-accounting-setup.entity.ts
export class DefaultAccountingSetup {
    accountingSetupId: number;
    cashReceived: string;
    cashPaid: string;
    purchaseInvoice: string;
    salesInvoice: string;
    washoutInvoiceGains: string;
    washoutInvoiceLoss: string;
    fxReval: string;
    defaultBankAccount: string;
    salesLedgerControlClientDebtors: string;
    purchaseLedgerControlClientCreditors: string;
    fxRevalaccount: string;
    suspenseAccountforWashoutSuspense: string;
    realisedPhysicalsPayable: string;
    realisedPhysicalsReceivable: string;
    vatAccountInputs: string;
    vatAccountOutputs: string;
    fxAccountGain: string;
    fxAccountLoss: string;
    cancellationGain: string;
    cancellationLoss: string;
    plClearanceYepAccount: string;
    balanceSheetClearanceYepAccount: string;
    bsReserveYepAccount: string;
    yepCostType: string;

    dealNominalAccount: string;
    settlementNominalAccount: string;
    yepDepartment: string;
}